added api to use concept search

+++ public-api/src/test/java/org/pmiops/workbench/cdr/dao/DbDomainDaoTest.java
@@ -125,7 +125,7 @@ public class DbDomainDaoTest {
     @Test
     public void findDomainTotals() throws Exception{
         final List<DbDomain> list=dao.findDomainTotals();
-        Assert.assertEquals(list.get(0).getCountValue(),25);
+        Assert.assertEquals((Long)list.get(0).getCountValue(),Long.valueOf(25));
         Assert.assertNotEquals(list,null);
     }
